Split checking and processing of parameters to allow for multiple fit calls
Multiple calls to fit cause the method to fail. The dist_measure string was overwritten with a function, causing later checks to fail. Alternative implementation would be to allow for checking whether one of the allowed functions is passed, this would also be consistent with the behaviour of check_random_state from scikit-learn.